reordered preprocessing operations for precision

+++ 5-yolo.py
@@ -321,9 +321,10 @@ class Yolo:
 
         for img in images:
             image_shapes.append((img.shape[0], img.shape[1]))
+            img = img.astype(np.float32)
             resized = cv2.resize(img, input_size,
                                  interpolation=cv2.INTER_CUBIC)
-            pimg = resized.astype(np.float32) / 255.0
+            pimg = resized / 255.0
             pimgs.append(pimg)
 
         pimgs = np.array(pimgs).reshape(-1, input_size[1], input_size[0], 3)
